% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{check_login}
\alias{check_login}
\title{Check Slack Login}
\usage{
check_login(team_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{session}{The shiny session object. In most situations you should use
the default.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} which returns a logical indicating whether the
user is logged in with proper API access.
}
\description{
Confirm that a user is logged into Slack.
}
