% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis.R
\name{.do_login}
\alias{.do_login}
\title{Perform the Login Via Slack}
\usage{
.do_login(site_url, team_id)
}
\arguments{
\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
A function that takes a request and returns a \code{\link[shiny:reexports]{shiny::tagList()}}that
sends the user to the proper Slack api URL in order to authenticate.
}
\description{
Perform the Login Via Slack
}
\keyword{internal}
