% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis.R
\name{.extract_full_url}
\alias{.extract_full_url}
\title{Keep Url Bits}
\usage{
.extract_full_url(request)
}
\arguments{
\item{request}{The shiny request object to parse.}
}
\value{
The url with query parameters included.
}
\description{
Keep Url Bits
}
\keyword{internal}
