% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis.R
\name{.extract_server_url}
\alias{.extract_server_url}
\title{Extract the server URL from the request}
\usage{
.extract_server_url(request)
}
\arguments{
\item{request}{A request object.}
}
\value{
A character with the url.
}
\description{
Extract the server URL from the request
}
\keyword{internal}
