% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_wrapper.R
\name{.parse_app_args}
\alias{.parse_app_args}
\title{Figure Out Args for shinyslack_app}
\usage{
.parse_app_args(options, site_url)
}
\arguments{
\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}
}
\value{
A list with elements \code{options} and \code{site_url}.
}
\description{
Figure Out Args for shinyslack_app
}
\keyword{internal}
