% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouath.R
\name{.parse_auth_code}
\alias{.parse_auth_code}
\title{Convert a Slack Authorization Code to a Token}
\usage{
.parse_auth_code(request, site_url, team_id)
}
\arguments{
\item{request}{The shiny request object to parse.}

\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
A \code{\link[shiny]{tagList}} that sets the cookie then reloads the
site.
}
\description{
Convert a Slack Authorization Code to a Token
}
\keyword{internal}
