% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouath.R
\name{.parse_ui}
\alias{.parse_ui}
\title{Prepare a Shiny UI for Display}
\usage{
.parse_ui(ui, request)
}
\arguments{
\item{ui}{A function defining the UI of a Shiny app, or a
\code{\link[shiny]{tagList}}.}

\item{request}{The shiny request object to parse.}
}
\value{
The parsed ui, as a \code{\link[shiny]{tagList}}.
}
\description{
Prepare a Shiny UI for Display
}
\keyword{internal}
