% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters Used in Various Functions}
\arguments{
\item{request}{The shiny request object to parse.}

\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{input}{The shiny input object.}
}
\description{
Parameters Used in Various Functions
}
\keyword{internal}
