% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{.shinyslack_decrypt}
\alias{.shinyslack_decrypt}
\title{Decrypt a String If Possible}
\usage{
.shinyslack_decrypt(string)
}
\arguments{
\item{string}{A length-1 character to encrypt (or decrypt).}
}
\value{
If \code{shinyslack_key} environment variable is set, the decrypted string
(or "bad_string" if decryption fails). Otherwise the original string is
returned.
}
\description{
Decrypt a String If Possible
}
\keyword{internal}
