% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{.slack_token_cookie_name}
\alias{.slack_token_cookie_name}
\title{The Name of the Slack Token Cookie}
\usage{
.slack_token_cookie_name(team_id)
}
\arguments{
\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
The name of the Slack token cookie
}
\description{
The Name of the Slack Token Cookie
}
\keyword{internal}
