% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis.R
\name{.update_site_url}
\alias{.update_site_url}
\title{Keep Url Bits}
\usage{
.update_site_url(site_url, request)
}
\arguments{
\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{request}{The shiny request object to parse.}
}
\value{
The site_url with query parameters included.
}
\description{
Keep Url Bits
}
\keyword{internal}
