% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouath.R
\name{.validate_cookie_token}
\alias{.validate_cookie_token}
\title{Make Sure a Cookie Token Works}
\usage{
.validate_cookie_token(cookie_token, team_id)
}
\arguments{
\item{cookie_token}{A character with the code used to authenticate the user.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
A logical indicating whether the token works for testing
authentication for this team.
}
\description{
Make Sure a Cookie Token Works
}
\keyword{internal}
