% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{include_js_cookie}
\alias{include_js_cookie}
\title{Add the js-cookie Javascript Library}
\usage{
include_js_cookie()
}
\value{
An html_dependency, which Shiny uses to add the js-cookie Javascript
library exactly once.
}
\description{
We only use the \code{set} method from this library for easier cookie
handling, but we make it easy to add the library in case you wish to
manipulate cookies.
}
