% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{oauth-extractors}
\alias{oauth-extractors}
\alias{.extract_url_token}
\alias{.extract_auth_code}
\title{Extract Oauth Objects from a Shiny Request}
\usage{
.extract_url_token(request)

.extract_auth_code(request)
}
\arguments{
\item{request}{The shiny request object to parse.}
}
\value{
The oauth object if it exists, or NULL.
}
\description{
Extract Oauth Objects from a Shiny Request
}
\keyword{internal}
