% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{request-predicates}
\alias{request-predicates}
\alias{.has_token}
\alias{.has_cookie_token}
\alias{.has_url_token}
\alias{.has_auth_code}
\title{Request Predicates}
\usage{
.has_token(request, team_id)

.has_cookie_token(request, team_id)

.has_url_token(request)

.has_auth_code(request)
}
\arguments{
\item{request}{The shiny request object to parse.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
Logical scalar.
}
\description{
These predicate functions check for the presence of various objects in the
request object.
}
\keyword{internal}
