% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyslack-package.R
\docType{package}
\name{shinyslack-package}
\alias{shinyslack}
\alias{shinyslack-package}
\title{shinyslack: Integrate Slack and Shiny}
\description{
Login to Shiny apps using Slack, and use Slack information in those apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/shinyslack}
  \item Report bugs at \url{https://github.com/r4ds/shinyslack/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
