% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get_shinyslack_team_id}
\alias{get_shinyslack_team_id}
\alias{set_shinyslack_team_id}
\title{Get the current team_id}
\usage{
get_shinyslack_team_id()

set_shinyslack_team_id(team_id)
}
\arguments{
\item{team_id}{The Slack team ID through which the user is being
authenticated.}
}
\value{
A string representing the team_id.
}
\description{
The \code{team_id} is set when an app is launched. In almost all cases, that value
is the one you will want for any instances of \code{team_id}.
}
\examples{
# If no app is active, the team_id will be a zero-length character vector.
get_shinyslack_team_id()

set_shinyslack_team_id("T123456")
get_shinyslack_team_id()
}
