% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_info}
\alias{user_info}
\title{Fetch Slack User Info}
\usage{
user_info(components = c("user_name", "user_id"))
}
\arguments{
\item{components}{A character vector of user components to include
(\code{"user_name"} and/or \code{"user_id"}).}
}
\value{
A \code{\link[shiny]{reactive}} with a named character vector.
}
\description{
Get information about the logged-in user from the Slack API.
}
