% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_wrapper.R
\name{.parse_app_args}
\alias{.parse_app_args}
\title{Figure Out Args for shinyslack_app}
\usage{
.parse_app_args(options)
}
\arguments{
\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}
}
\value{
The parsed \code{options}.
}
\description{
Figure Out Args for shinyslack_app
}
\keyword{internal}
