% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis.R
\name{.parse_auth_code}
\alias{.parse_auth_code}
\title{Convert a Slack Authorization Code to a Token}
\usage{
.parse_auth_code(site_url, team_id, expiration)
}
\arguments{
\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{expiration}{Days after which the cookie should expire.}
}
\value{
A function that takes a request and returns a \code{\link[shiny:reexports]{shiny::tagList()}} that
sets the cookie then reloads the site.
}
\description{
Convert a Slack Authorization Code to a Token
}
\keyword{internal}
