% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters Used in Various Functions}
\arguments{
\item{request}{The shiny request object to parse.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}

\item{session}{The shiny session object. The default
\code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}} is likely always sufficient outside of
tests.}

\item{shinyslack_key}{(optional) A key to use to encrypt the string. If not
set, the string is returned unencrypted.}

\item{slack_api_key}{The Slack API key to use. The default value should
likely always be used outside of tests.}
}
\description{
Parameters Used in Various Functions
}
\keyword{internal}
