% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters Used in Various Functions}
\arguments{
\item{request}{The shiny request object to parse.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}
}
\description{
Parameters Used in Various Functions
}
\keyword{internal}
