% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{.shinyslack_decrypt}
\alias{.shinyslack_decrypt}
\title{Decrypt a String If Possible}
\usage{
.shinyslack_decrypt(string, shinyslack_key = Sys.getenv("SHINYSLACK_KEY"))
}
\arguments{
\item{shinyslack_key}{(optional) A key to use to encrypt the string. If not
set, the string is returned unencrypted.}
}
\value{
If \code{shinyslack_key} is non-empty, the decrypted string (or
"bad_string" if decryption fails). Otherwise the original string is
returned.
}
\description{
Decrypt a String If Possible
}
\keyword{internal}
