% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{slack_shiny_ui}
\alias{slack_shiny_ui}
\title{Require Slack login to a Shiny app}
\usage{
slack_shiny_ui(
  ui,
  team_id,
  expiration = 90,
  shinyslack_key = Sys.getenv("SHINYSLACK_KEY")
)
}
\arguments{
\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{expiration}{Days after which the cookie should expire. To remove an
HttpOnly cookie, send a negative value for this attribute.}

\item{shinyslack_key}{(optional) A key to use to encrypt the string. If not
set, the string is returned unencrypted.}
}
\value{
A function defining the UI of a Shiny app (either with login or
without).
}
\description{
This is a function factory that wraps a Shiny ui. If the user does not have a
cookie for that site, they are prompted to login. Once they have a cookie,
the UI displays as normal. #5
}
