% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_wrapper.R
\name{slack_shiny_ui}
\alias{slack_shiny_ui}
\title{Require Slack login to a Shiny app}
\usage{
slack_shiny_ui(ui, team_id, site_url, expiration = 90)
}
\arguments{
\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}

\item{team_id}{The Slack team ID through which the user is being
authenticated.}

\item{site_url}{The URL of the Shiny app. This is used to return the user to
the app from the Slack login.}

\item{expiration}{Days after which the user's login cookie should expire.}
}
\value{
A function defining the UI of a Shiny app (either with login or
without).
}
\description{
This is a function factory that wraps a Shiny ui. If the user does not have a
cookie for that site, they are prompted to login. Once they have a cookie,
the UI displays as normal.
}
