% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_info}
\alias{user_info}
\title{Fetch Slack User Info}
\usage{
user_info(
  components = c("user_id", "real_name", "display_name", "pronouns", "user_name")
)
}
\arguments{
\item{components}{A character vector of user components to include. Current
options are:
\itemize{
\item user_id: The ID used to uniquely identify this user on this Slack team.
\item real_name: The full name of this user as entered in their profile.
\item display_name: The name that the user has chosen to display to other users in Slack.
\item pronouns: The pronouns set by this user, if any.
\item user_name: You probably do not want this. It is a legacy piece of information.
}}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with a named character vector.
}
\description{
Get information about the logged-in user from the Slack API.
}
